// front/src/styles/components/Stampcards/StampcardSpot.module.scss.d.ts
export type Styles = {
  address: string
  blue: string
  buttons_wrapper: string
  card_wrapper: string
  comment: string
  description_wrapper: string
  green: string
  image: string
  image_wrapper: string
  name: string
  name_wrapper: string
  number: string
  pink: string
  slide_wrapper: string
  swiper: string
  wrapper: string
  yellow: string
}

export type ClassNames = keyof Styles

declare const styles: Styles

export default styles